Normalize helpers and handle JSON
